BAN-5 refactor: change stories
