+++ src/shared/ui/index.ts
@@ -1,4 +1,4 @@
 export {Button} from './button/Button'
 export {Textarea} from './textarea/Textarea'
 export {Pagination} from './pagination/Pagination'
-export {Input} from './input/Input'
+export {Input,InputSlot,type InputProps} from './input/Input'
